Simple e-2-e test.
